added first project to portflio
added a .aboutProject div for text that gives details about the project. Details will include: name, languages and purpose of project.
Also learned how to position the button so it's in the middle of the image. Resized the continers.

next time I want to widen the portfolio section so the img containers flex